include: Move array_map.hh to util/internal/
And to internal namespace either. And adjust the callers.

Callers could use the std::array<> directly, but they need C99-style
designated initializers for convenience.
