+++ src/seastar.cc
@@ -166,7 +166,6 @@ export module seastar;
 #include <seastar/core/align.hh>
 #include <seastar/core/aligned_buffer.hh>
 #include <seastar/core/app-template.hh>
-#include <seastar/core/array_map.hh>
 #include <seastar/core/bitops.hh>
 #include <seastar/core/bitset-iter.hh>
 #include <seastar/core/byteorder.hh>
