+++ src/util/log.cc
@@ -52,7 +52,7 @@ module seastar;
 #include <seastar/util/log.hh>
 #include <seastar/util/log-cli.hh>
 
-#include <seastar/core/array_map.hh>
+#include <seastar/util/internal/array_map.hh>
 #include <seastar/core/reactor.hh>
 #include <seastar/core/future.hh>
 #include <seastar/core/print.hh>
@@ -87,7 +87,7 @@ template <> struct formatter<wrapped_log_level> {
 
     template <typename FormatContext>
     auto format(wrapped_log_level wll, FormatContext& ctx) const {
-        static seastar::array_map<seastar::sstring, nr_levels> text = {
+        static seastar::internal::array_map<seastar::sstring, nr_levels> text = {
             { int(log_level::debug), "DEBUG" },
             { int(log_level::info),  "INFO " },
             { int(log_level::trace), "TRACE" },
@@ -96,7 +96,7 @@ template <> struct formatter<wrapped_log_level> {
         };
         int index = static_cast<int>(wll.level);
         std::string_view name = text[index];
-        static seastar::array_map<text_style, nr_levels> style = {
+        static seastar::internal::array_map<text_style, nr_levels> style = {
             { int(log_level::debug), fg(terminal_color::green)  },
             { int(log_level::info),  fg(terminal_color::white)  },
             { int(log_level::trace), fg(terminal_color::blue)   },
@@ -364,7 +364,7 @@ logger::do_log(log_level level, log_writer& writer) {
         auto it = buf.back_insert_begin();
         it = print_once(it);
         *it = '\0';
-        static array_map<int, 20> level_map = {
+        static internal::array_map<int, 20> level_map = {
                 { int(log_level::debug), LOG_DEBUG },
                 { int(log_level::info), LOG_INFO },
                 { int(log_level::trace), LOG_DEBUG },  // no LOG_TRACE
